fix: update methods to add and find vaults based on user id
This commit fixes the `#add_vault` and `#find_vault` methods of the
`DatabasePersistence` class.

Previously, the username was used to find the user id, which can result
in `nil` values if a user can't be found. It is now updated to use the
user id instead to add or find a vault entry associated with the user
id.