feat: find user and authenticate password at login
This commit adds the `::login` method that returns a `User` object if a
user of the given username is found, and verifies if the given password
matches the hashed password.

It also adds the `::authenticate` method
that uses `BCrypt` to check if the hashed password matches the given
password. It returns the `User` instance if the passwords match, and
`false` otherwise.