feat: set attributes if given when instantiating `User` object
This commit adds the `#set_attributes` method that is invoked when
instantiating a new `User` object with `::new`. It initializes the
instance variables to the given values if any optional arguments are
passed to `::new`.

`::create_user_from_tuple` performs a similar function to
`#set_attributes`, so it is removed. Instead, a new `User` instance is
instantiated and passed the tuple as an argument in the
`::find_by_username` method.

If no arguments are passed to `::new`, a new `User` object is
instantiated without any instance variables initialized.