refactor: move validation of user sign-up credentials to `User` class
This commit moves the logic from `::error_for_new_username` and
`::error_for_new_password` from the Validation module into the `User`
class, which are `#username_validation` and `#password_validation`
respectively. These instance methods check the username and passwords
within the `User` instance and, if applicable, adds an error message to
the `@errors` variable.

The `#validate` method is able to invoke the appropriate validation
methods by passing it the attribute name as a symbol.