feat: add a user to the database upon successful sign-up
This commit adds a `::create` method that checks if the given username
and passwords are valid. If they are not valid, a SignupError is
returned with the error messages returned by the validation methods. If
no error is raised, the username and password hash are added to the
`users` table and the `User` instance is returned.