feat: set attributes if given when instantiating `Vault` object
This commit adds the `#set_attributes` method that is invoked when
instantiating a new `Vault` object with `::new`. It initializes the
instance variables to the given values if any optional arguments are
passed to `::new`.

If no arguments are passed to `::new`, a new `Vault` object is
instantiated without any instance variables initialized.