test: move tests for `User` class to a separate file
This commit moves the tests to add and find users to a separate test
file.

It also updates the test to invoke `:add` and `::find_by_username` to
add and find users from the `users` table.`#authenticate` is used to
verify the password rather than comparing it with the password hash
directly.

It also splits the tests for verifying the return value when querying
for a valid vs non-existent user.