test: move tests for `DatabaseAccessor` class to a separate file
This commit moves the test verify resetting the database to a separate
file. It also updates the test to invoke methods from the `User` and
`Vault` class.

The getter method for the `@id` variable is also defined in the `User`
class so that it can be referenced in the test.