refactor: update sign-up process to invoke `User` and `Vault` methods
This commit updates the application file to invoke the `User::create`
and `Vault:add` methods to create a user and vault entity upon sign-up.
If the given username and/or passwords are invalid, a `SignupError` is
raised, and the error message is stored in the session hash.

Validation is done within the `User` class, so the methods in the
`Validation` module can be deleted. The respective test file is also
updated to reflect the new changes.