refactor: update sign-up process to invoke `User` methods
This commit updates the application file to invoke the `User::login`
method to validate the username and password upon sign-in. If the given
username and/or password is invalid, a `LoginError` is raised.

Instead of the previous specific error messages given for an invalid
username or password, a generic message is given to prevent malicious
attackers from knowing that the username is valid.

Validation is done with the `User` class, so the `Validation` module is
deleted. The respective test file is also updated to reflect the new
changes.