style: rename and move application file
This commit renames the `password_manager.rb` file to `app.rb` and moves
it out of the `lib` directory. This differentiates the application file
from other ruby files. It also moves the `public` and `views` directory
out of the `lib` directory. It also updates the paths in `app.rb` that
load the `DatabaseAccessor`, `User`, and `Vault` classes.

It also renames the test file to `app_test.rb`. It updates the path to
load the application file, and the class name for the test.