feat: add error types and messages to `Error` object
This commit adds a method that modifies the `@errors` hash in the
`Error` instance by defining error types as hash keys, and an array of
error messages as hash values. It also adds a method that returns a copy
of the `@errors` hash.

It updates the test file to verify that the error type and message are
added.