feat: update `::login` method to use `Error` instance
This commit updates the `User::login` method to implement the
`Error#add` method to save the error type and message in `@errors` in
the event that the login credentials are invalid.

The updated method returns a new `User` object with the updated `Error`
instance referenced by `@errors` if no valid user is found.

It also removes the `LoginError` class.