refactor: store `id` and `username` separately in session
This commit adds a `#save_user_info_in_session` method that assigns the
user `id` and `username` to separate keys in the session hash. This
makes retrieval of either value easier. It updates the test file for the
main application to reflect the changes.

This commit deletes the redundant `#session_hash` method from the `User`
class.