feat: add `credentials` table
This commit updates the schema to add a `credentials` table that
includes the columns:
- id (primary key, auto-increment)
- user_id (not null, foreign key that references `id` in the `users`
table)
- name (not null)
- username (not null)
- encrypted password
- iv
- notes

There is a UNIQUE constraint for (`name`, `username`) such that the
name-username combination cannot be repeated. This change allows users
to store credentials to the services they are using. Encrypted password
and iv are not required since some services may involve single sign-on.