feat: define `Credentials` class
This commit defines a new `Credentials` class which will contain
information about the credentials stored by the user. It checks if
a cipher key is available upon instantiation, and if not, generates
one on the user's local machine.

This commit also adds tests that verifies a new cipher key is
created on the user's local machine if it does not already exist.