feat: update username validation
This commit updates `User#username_validation` to verify that the given
username has between 2 and 36 characters. Logic validating whether the
username is already in the database is moved to `User#username_unique?`.

This commit also updates the test file to verify that an error message
is displayed if the username is not of the correct length.