feat: update password validation
This commit updates `User#password_validation` to verify that a given
password has between 6 - 125 characters and that it has at least one
letter and one number. Logic validating if the passwords match is moved
to `User#passwords_match?`.

This commit also updates the test file so that all tests not verifying
password validation logic have passwords of the required length and
characters. It also adds additional tests to verify that an error
message will be display if the password length is incorrect or if the
password lacks the required characters.