feat: add name validation for credentials
This commit adds a `#name_validation` method in the `Credentials` class
to check that the name has between 1 and 64 characters. An `Error`
instance is added to the `Credentials` instance if the name is not
valid.

It also adds a test that verifies that the `Error` instance is added
with the appropriate error message if the name is not valid.