feat: add username validation for credentials
This commit adds a `#username_validation` method in the `Credentials`
class to check that the username has between 2 and 256 characters. An
`Error` instance is added to the `Credentials` instance if the username
is not valid.

It also adds a test that verifies that the `Error` instance is added
with the appropriate error message if the username is not valid.