feat: add validated credentials to `credentials` table
This commit includes the `Credentials#add` method which inserts a new
row of data from the `Credentials` instance into the `credentials`
table. It also includes `CHECK` constraints to the `credentials` table
schema to validate the data being inserted.

This commit also adds tests to verify that a `PG::CheckViolation` error
is raised if attempting to insert credentials with an invalid name or
username.