feat: find credentials by the name and username
This commit adds the `Credentials::find_by_name_and_username` method
that returns a `Credential` instance if a record is found in
the `credentials` table with the given name and username. If not, it
returns `nil`.

This commit also adds additional tests verifying that a valid
`Credentials` instance is returned if the record is found, and `nil` is
returned if there is no matching record.