feat: update `DatabaseAccessor::reset` to reset `credentials`
This commit updates the `DatabaseAccessor::reset` method to delete all
data from the `credentials` table and reset the `id` sequence to `1`.

It also updates the related test file to verify that all records in the
`credentials` table are deleted upon `::reset` invocation and that the
`id` sequence is reset to `1`.