feat: validate that the credentials given is unique
This commit adds a `Credentials#unique?` method that queries the
database for a record with a matching `name` and `username`. If the
record exists and the `id` does not match the current `Credentials`
instance, the `@errors` variable is updated and the method returns
`false`. Otherwise, `#unique?` returns `true`.

This commit also adds an additional test that verifies that `#unique?`
returns true if no matching record is found or if the matching record
has the same `id` as the current instance. Otherwise, it returns `false`
and the value of `@errors` is updated.