feat: encrypt the password
This commit adds the `Credentials#encrypt_password` method that
generates an iv for each password encryption. Using the key stored on
the local machine, the password is encrypted with the generated iv. The
`@password` variable is removed and the encrypted password is assigned
to `@encrypted_password`.

This commit also updates the `Credentials#add` method to encrypt the
password for the instance if it exists. An additional test is also added
to verify that the `@password` variable is removed upon encryption, and
a `@encrypted_password` variable is initialized. The value of
`@encrypted_password` does not match the initial value of `@password`.