feat: encode iv and encrypted password in Base64
This commit requires the `Base64` module and adds a
`Credentials#encode_iv_and_password` method which encodes the values of
`@iv` and `@encrypted_password` using a Base64 representation. Inserting
the iv and encrypted password into the `credentials` table raises a
`PG::CharacterNotInRepertoire` error. Encoding is done to represent the
values in a string format that can be stored in the database.