feat: create and add credentials to database if valid
This commit adds a `Credentials::create` method which invokes the
`::new` method to instantiate a `Credentials` instance with the given
attributes. It validates the attributes prior to adding the object to
the database. If the given credentials is not unique or any of the
attributes are invalid, the object is not added to the database.
Instead, it is returned with an updated `Errors` instance.

This commit also adds tests that verify the above behavior.