feat: add credentials upon form submission
This commit adds a new `POST` route `/:username/passwords` which
instantiates and validates a new `Credentials` object.

If the given name and username is unique and valid, the given password
(if it exists) is encrypted, and all the data is added to the
`credentials` table and the user is redirected to the user's dashboard.

Otherwise, the data is not added, and the `Errors` instance referenced
by `@errors` is updated instead. The form to add credentials is
re-rendered and the error messages are printed.

This commit also adds tests that verify the above behaviors.