refactor: change `User::add` to `User#add`
This commit changes the `User::add` class method to the `User#add`
instance method. This standardizes the `#add` functionality so it
functions similarly to the `#add` method of the `Credentials` class. An
additional `#hash_password` method is also defined to remove the
`@password` and `@password_confirmation` variables and initialize the
`@password_hash` variable to reference the hash generated by `BCrypt`.

This commit also updates the test file for the `User` class.