refactor: return `id` upon adding user to database
This commit updates the SQL statement in the `User#add` method to return
the value in the `id` column upon inserting the data into the `users`
table. The `@id` variable is initialized and references the returned
value.

Instead of executing a query to find the matching record, the
`User` object with the updated `@id` is now returned in the `::create`
method.