feat: update schema to mirror validations in `User` class
This commit includes `CHECK` constraints to ensure that the username
inserted is between 2-36 characters and only contains alphanumeric
characters. It adds a test to verify this behavior.

This commit removes the `UNIQUE` constraint for the
`username`-`password_hash` pair since `username` already has a `UNIQUE`
constraint. It also includes additional whitespace for the `UNIQUE`
constraint of `credentials` table.