feat: find and display all associated credentials
This commit invokes the `Credentials::find_all_by_user_id` method in the
`GET "/passwords"` route to retrieve all credentials stored with the
associated user id. It updates the ERB template to display a HTML table
with the data for each set of stored credentials if any credentials are
found.

This commit also defines getter methods for `@notes`, `@created_at` and
`@updated_at` variables to reference their values in the table.